chore: Remove unnecessary print statement in main function
Removed the unnecessary print statement in the `main` function. The print statement was originally added for debugging purposes but is no longer needed in the final version of the code. By removing this print statement, we improve the overall cleanliness and readability of the code. This change does not affect the functionality of the program in any way.

The print statement was located at the beginning of the `main` function and was printing out a variable that was no longer being used. This variable was previously used for tracking the execution flow but has since been replaced with a more efficient and reliable method. Therefore, the print statement became redundant and was removed to avoid confusion