feat: Move import statement for `openai` module
The import statement for the `openai` module was moved to a different location in the code. This change was made to improve the organization and readability of the codebase.

Additionally, an import statement for the `logging_utils` module was added. This module provides utility functions for logging, which will be used to enhance the logging capabilities of the code.

In order to reduce unnecessary output, the `print(ini_path)` statement was commented out. This statement was previously used for debugging purposes but is no longer needed in the final version of the code.

Furthermore, the `logger